refine sign_up
