+++ src/auth/sign.rs
@@ -28,6 +28,32 @@ pub async fn sign_in_email(
 
 }
 
+pub async fn sign_up_email(
+    email: &str,
+    password: &str,
+    return_secure_token: bool
+) -> Result<SignUpResponse, AuthError> {
+    let url = format!(
+        "https://identitytoolkit.googleapis.com/v1/accounts:signUp?key={}",
+        "AIzaSyBvAE59iedRLnTKZYR1XRLw_4ozM8sx80k",
+    );
+
+    let request = serde_json::json!({
+        "email": email,
+        "password": password,
+        "return_secure_token": return_secure_token,
+    });
+
+    let auth_result = auth(&request, &url).await;
+
+    if let Err(AuthError::Auth(e)) = auth_result {
+        return Err(AuthError::SignUp(e));
+    }
+
+    auth_result
+
+}
+
 pub async fn auth<T: DeserializeOwned>(request: & serde_json::Value, url: &str) -> Result<T, AuthError> {
 
     let client = awc::Client::new();
